+++ lib/dropdown/drop_down_view_wrapper.dart
@@ -2,26 +2,20 @@ import 'package:flutter/material.dart';
 
 import 'drop_down_typedef.dart';
 
-class DropDownViewWrapper extends DropDownViewStatefulWidget {
+class DropDownViewWrapper extends DropDownViewStatelessWidget {
   final double height;
   final Widget child;
 
-  const DropDownViewWrapper({
-    super.key,
+  DropDownViewWrapper({
     required this.height,
     required this.child,
   });
 
   @override
-  State<DropDownViewWrapper> createState() => _DropDownViewWrapperState();
-
-  @override
   double get actualHeight => height;
-}
 
-class _DropDownViewWrapperState extends State<DropDownViewWrapper> {
   @override
   Widget build(BuildContext context) {
-    return widget.child;
+    return child;
   }
 }